# docs/openapi-schema.yml
openapi: 3.0.2
info:
  title: Studygram API
  version: "1.0.0"
  description: API платформы для создания, поиска и сохранения учебных материалов.
paths:
...
  /api/materials/{id}/:
    get:
      operationId: Получение учебного материала
      description: Получение учебного материала по id.
      parameters:
        - name: id
          in: path
          required: true
          description: Уникальный id материала.
          schema:
            type: integer
      responses:
        "200":
          description: Данные учебного материала.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
        "404":
          $ref: "#/components/responses/NotFound"
      tags:
        - Материалы

    patch:
      operationId: Обновление учебного материала
      description: Обновление учебного материала. Доступно только автору.
      security:
        - Token: []
      parameters:
        - name: id
          in: path
          required: true
          description: Уникальный id материала.
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialUpdate"
      responses:
        "200":
          description: Материал успешно обновлён.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationError"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"
      tags:
        - Материалы

    delete:
      operationId: Удаление учебного материала
      description: Удаление учебного материала. Доступно только автору.
      security:
        - Token: []
      parameters:
        - name: id
          in: path
          required: true
          description: Уникальный id материала.
          schema:
            type: integer
      responses:
        "204":
          description: Материал успешно удалён.
        "401":
          $ref: "#/components/responses/AuthenticationError"
        "403":
          $ref: "#/components/responses/PermissionDenied"
        "404":
          $ref: "#/components/responses/NotFound"
      tags:
        - Материалы
...
components:
  schemas:
    User:
      description: Пользователь платформы Studygram.
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          description: Адрес электронной почты.
        id:
          type: integer
          readOnly: true
        username:
          type: string
          maxLength: 150
          description: Уникальный username.
          example: study_user
        first_name:
          type: string
          maxLength: 150
          description: Имя.
          example: Иван
        last_name:
          type: string
          maxLength: 150
          description: Фамилия.
          example: Петров
        is_subscribed:
          type: boolean
          readOnly: true
          description: Подписан ли текущий пользователь на этого автора.
          example: false
        avatar:
          type: string
          nullable: true
          format: uri
          description: Ссылка на аватар.
          example: http://studygram.example.org/media/users/avatar.png
      required:
        - username
...
    Topic:
      description: Тема учебного материала.
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 64
          description: Название темы.
          example: Python
        slug:
          type: string
          maxLength: 64
          description: Уникальный slug темы.
          example: python
      required:
        - name
        - slug
    Material:
      description: Учебный материал.
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        topics:
          type: array
          readOnly: true
          description: Список тем материала.
          items:
            $ref: "#/components/schemas/Topic"
        author:
          readOnly: true
          $ref: "#/components/schemas/User"
        is_favorited:
          type: boolean
          readOnly: true
          description: Добавлен ли материал в избранное.
        title:
          type: string
          maxLength: 256
          description: Название материала.
          example: Основы Django REST Framework
        description:
          type: string
          description: Краткое описание материала.
          example: Введение в создание REST API на Django.
        content_type:
          type: string
          description: Тип учебного материала.
          example: article
          enum:
            - article
            - video
            - course
            - book
            - documentation
            - other
        link:
          type: string
          format: uri
          description: Ссылка на учебный материал.
          example: https://www.django-rest-framework.org/
        estimated_time:
          type: integer
          nullable: true
          description: Примерное время изучения в минутах.
          example: 60
        level:
          type: string
          nullable: true
          description: Уровень сложности.
          example: beginner
          enum:
            - beginner
            - intermediate
            - advanced
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
...
    MaterialUpdate:
      description: Обновление учебного материала.
      type: object
      properties:
        topics:
          description: Список id тем.
          type: array
          example: [1, 2]
          items:
            type: integer
        title:
          type: string
          maxLength: 256
          description: Название материала.
        description:
          type: string
          description: Описание материала.
        content_type:
          type: string
          description: Тип материала.
        link:
          type: string
          format: uri
          description: Ссылка на материал.
        estimated_time:
          type: integer
          nullable: true
          description: Примерное время изучения в минутах.
        level:
          type: string
          nullable: true
          description: Уровень сложности.
...
    ValidationError:
      description: Ошибки валидации DRF.
      type: object
      properties:
        field_name:
          type: array
          items:
            type: string
          example:
            - Обязательное поле.
...
    AuthenticationError:
      description: Пользователь не авторизован.
      type: object
      properties:
        detail:
          type: string
          example: Учетные данные не были предоставлены.
    PermissionDenied:
      description: Недостаточно прав.
      type: object
      properties:
        detail:
          type: string
          example: У вас недостаточно прав для выполнения данного действия.
    NotFound:
      description: Объект не найден.
      type: object
      properties:
        detail:
          type: string
          example: Страница не найдена.
  responses:
    ValidationError:
      description: Ошибки валидации.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
    AuthenticationError:
      description: Пользователь не авторизован.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AuthenticationError"
    PermissionDenied:
      description: Недостаточно прав.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PermissionDenied"
    NotFound:
      description: Объект не найден.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/NotFound"
...
  securitySchemes:
    Token:
      description: 'Авторизация по токену. Все запросы от имени пользователя должны выполняться с заголовком "Authorization: Token TOKENVALUE".'
      type: http
      scheme: token